+++ sdk/storage/Azure.Storage.Common/src/Shared/AesGcm/SafeBCryptHandle.cs
@@ -1,0 +1,34 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+using System;
+using System.Runtime.InteropServices;
+
+namespace Azure.Storage
+{
+    /// <summary>
+    /// From
+    /// https://github.com/dotnet/runtime/blob/main/src/libraries/Common/src/Microsoft/Win32/SafeHandles/SafeBCryptHandle.cs
+    ///
+    /// SHOULD NOT BE CHANGED WITHOUT COORDINATING WITH BCL TEAM
+    /// </summary>
+    internal abstract class SafeBCryptHandle : SafeHandle, IDisposable
+    {
+        protected SafeBCryptHandle()
+            : base(IntPtr.Zero, true)
+        {
+        }
+
+        public sealed override bool IsInvalid
+        {
+            get
+            {
+                return handle == IntPtr.Zero;
+            }
+        }
+
+        protected abstract override bool ReleaseHandle();
+    }
+}
